Udated POST /movies and PUT, POST /movies/{movie_id} enpoints

+++ project/routers/movies.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter
 from fastapi import HTTPException
+from fastapi import Depends
 
 from typing import List
 
@@ -7,6 +8,9 @@ from ..schemas import MovieRequestModel
 from ..schemas import MovieResponseModel
 
 from ..database import Movie
+from ..database import User
+
+from ..common import get_current_user
 
 
 router = APIRouter(prefix='/movies')
@@ -29,21 +33,26 @@ async def get_movie(movie_id: int):
     return movie
 
 @router.post('', response_model=MovieResponseModel)
-async def create_movie(movie: MovieRequestModel):
+async def create_movie(movie: MovieRequestModel, user: User = Depends(get_current_user)):
 
     movie = Movie.create(
-        title = movie.title
+        title = movie.title,
+        user_id = user.id
     )
 
     return movie
 
 @router.put('/{movie_id}', response_model=MovieResponseModel)
-async def update_movie(movie_id: int, movie_request: MovieRequestModel):
+async def update_movie(movie_id: int, movie_request: MovieRequestModel, user: User = Depends(get_current_user)):
 
     movie = Movie.select().where(Movie.id == movie_id).first()
 
     if movie is None:
         raise HTTPException(status_code=404, detail="Movie not found")
+    
+    if movie.user_id != user.id:
+        raise HTTPException(status_code=403, detail="You are not allowed to update this movie")
+
 
     movie.title = movie_request.title
     movie.save()
@@ -51,12 +60,15 @@ async def update_movie(movie_id: int, movie_request: MovieRequestModel):
     return movie
 
 @router.delete('/{movie_id}', response_model=MovieResponseModel)
-async def delete_movie(movie_id: int):
+async def delete_movie(movie_id: int, user: User = Depends(get_current_user)):
 
     movie = Movie.select().where(Movie.id == movie_id).first()
 
     if movie is None:
         raise HTTPException(status_code=404, detail="Movie not found")
+    
+    if movie.user_id != user.id:
+        raise HTTPException(status_code=403, detail="You are not allowed to delete this movie")
 
     movie.delete_instance()
